Remove all warnings

+++ kernel/util.c
@@ -1,5 +1,6 @@
 #include "util.h"
 #include <stdint.h>
+#include <stddef.h>
 
 int atoi(char s[]) {
     int n = 0;
@@ -94,8 +95,8 @@ int strcmp(char *str1, char *str2) {
     return str1[i] - str2[i];
 }
 
-char *strncopy(char *destination, char *source, uint32_t maximum) {
-  int i;
+char *strncopy(char *destination, char *source, size_t maximum) {
+  size_t i;
   for (i = 0; source[i] != '\0' && i < maximum; i++) {
     destination[i] = source[i];
   }
